Take the last Element slot out of ComposerBar
Badges were the other untyped hole. Both surfaces already rendered the
same WorkspaceBadges through it, so the slot bought nothing but the
chance to drift again; ComposerBar takes the four fields and the optional
create-worktree handler and renders the badges itself. The chat wrapper
that existed only to read those fields off Chat goes with it.

ComposerBar now has no Option<Element> at all. Both composers hand it
data and get the same thing drawn from it, which is the property the
duplication kept breaking.
