+++ crates/page/vmux_chat/src/page/composer.rs
@@ -1,14 +1,13 @@
 use self::menu::{CommandMenu, MediaMenu, ResumeMenu};
 use self::options::{ChatComposerMenus, ChatModelMenu};
-use self::workspace::WorkspaceBadges;
 use super::approval::ChoiceList;
 use super::keys::ChatKeys;
 use super::state::Chat;
 use super::transcript::QueuedPrompts;
-use crate::event::{ChatPasteMedia, ChatPickFiles};
+use crate::event::{ChatCreateWorktree, ChatPasteMedia, ChatPickFiles};
 use dioxus::prelude::*;
 use vmux_ui::agent_accent::agent_accent;
-use vmux_ui::components::composer::PromptComposer;
+use vmux_ui::components::composer::{PROMPT_INPUT_ID, PromptComposer, focus_prompt_end};
 use vmux_ui::components::composer_bar::ComposerBar;
 use vmux_ui::hooks::send;
 use vmux_ui::i18n::translate;
@@ -79,6 +78,8 @@ fn ChatComposer(chat: Chat) -> Element {
 
 #[component]
 fn ComposerFooter(chat: Chat) -> Element {
+    let context = (chat.slash.composer_context)();
+    let offers_worktree = !context.is_worktree && context.can_manage_workspace;
     rsx! {
         ComposerBar {
             menu: chat.menu,
@@ -86,9 +87,15 @@ fn ComposerFooter(chat: Chat) -> Element {
             effort: chat.effort_chip(),
             project: chat.project_chip(),
             branch: chat.branch_chip(),
-            badges: Some(rsx! {
-                WorkspaceBadges { chat }
-            }),
+            is_git_repo: context.is_git_repo,
+            workspace_known: context.workspace_selected,
+            uncommitted: context.uncommitted,
+            ahead: context.ahead,
+            on_create_worktree: offers_worktree
+                .then(|| EventHandler::new(move |()| {
+                    let _ = send(&ChatCreateWorktree);
+                    focus_prompt_end(PROMPT_INPUT_ID);
+                })),
             status: chat.status(),
             active_subagents: (chat.activity_counts)().0,
             active_tasks: (chat.activity_counts)().1,
@@ -99,4 +106,3 @@ fn ComposerFooter(chat: Chat) -> Element {
 
 mod menu;
 mod options;
-mod workspace;
